requirements.txt
confluent-kafka==2.1.1
Django==4.2.2
django-ckeditor==6.6.1
django-cors-headers==4.1.0
djangorestframework==3.14.0
drf-yasg==1.21.6
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
Pillow==9.5.0